Rename from "_zip" to "_compressed"
