+++ account_financial_report_compressed/report/templates/account_report_partners_ledger.mako
