+++ account_financial_report_compressed/wizard/partners_ledger_wizard.py
@@ -18,25 +18,26 @@
 #    along with this program.  If not, see <http://www.gnu.org/licenses/>.
 #
 ##############################################################################
-{
-    'name': 'Add ZIP multiple export to ledger reports',
-    'version': '0.1',
-    'license': 'AGPL-3',
-    'author': 'Camptocamp',
-    'category': 'Generic Modules/Accounting',
-    'description': """
-        This module add a multiple export option for ledgers
-        (partners and general), in order to export one PDF file
-        per account. This is useful if the customer has many lines.
-    """,
-    'depends': ['account_financial_report_webkit'],
-    'demo_xml': [],
-    'init_xml': [],
-    'update_xml': [
-        'report.xml',
-        'wizard/general_ledger_wizard_view.xml',
-        'wizard/partners_ledger_wizard_view.xml',
-    ],
-    'active': False,
-    'installable': True,
-}
+from openerp.osv import orm
+
+
+class AccountReportPartnerLedgerWizard(orm.TransientModel):
+    _inherit = "partners.ledger.webkit"
+
+    def compressed_export(self, cr, uid, ids, context=None):
+        return self.check_report(cr, uid, ids, context=context)
+
+    def _print_report(self, cr, uid, ids, data, context=None):
+        context = context or {}
+        if context.get('compressed_export'):
+            data = self.pre_print_report(cr, uid, ids, data, context=context)
+            return {
+                'type': 'ir.actions.report.xml',
+                'report_name':
+                'account.account_report_partners_ledger_compressed',
+                'datas': data
+            }
+        else:
+            return super(AccountReportPartnerLedgerWizard, self)._print_report(
+                cr, uid, ids, data, context=context
+            )
