+++ account_financial_report_compressed/wizard/partners_ledger_wizard_view.xml
@@ -2,14 +2,14 @@
 <openerp>
   <data>
 
-      <record id="partners_ledger_webkit_zip_wizard_view" model="ir.ui.view">
-          <field name="name">partners.ledger.webkit.zip</field>
+      <record id="partners_ledger_compressed_wizard_view" model="ir.ui.view">
+          <field name="name">partners.ledger.compressed</field>
           <field name="model">partners.ledger.webkit</field>
           <field name="type">form</field>
           <field name="inherit_id" ref="account_financial_report_webkit.account_partner_ledger_view_webkit"/>
           <field name="arch" type="xml">
             <button string="Print" position="after">
-                <button icon="gtk-print" name="zip_export" string="Print (multiple files)" type="object" context="{'zip_export':1}" colspan="2"/>
+                <button icon="gtk-print" name="compressed_export" string="Print (multiple files)" type="object" context="{'compressed_export':1}" colspan="2"/>
             </button>
           </field>
       </record>
